requirements.txt
numpy==2.3.4
scipy==1.16.2
netCDF4==1.7.3
matplotlib==3.10.7
pillow